[Decomp][Unit] Cohesion split: extract power/resource accessors to UnitPower.cpp
